+++ src/app/about-us/components/mission-vision/mission-vision.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 import hugging from "/public/images/people-hugging-together-wide.webp";
 
-import { AboutUs } from "../about-us/about-us";
+import { AboutUs } from "../../../../components/about-us/about-us";
 import styles from "./mission-vision.module.scss";
 
 export function MissionVision() {
