+++ src/app/author/[id]/components/author-hero/author-hero.tsx
