+++ src/app/author/[id]/page.tsx
@@ -1,6 +1,7 @@
-import AuthorHero from "@/components/author-hero/author-hero";
-import { AuthorPosts } from "@/components/author-posts/author-posts";
 import { AUTHORS } from "@/data/authors";
+
+import AuthorHero from "./components/author-hero/author-hero";
+import { AuthorPosts } from "./components/author-posts/author-posts";
 
 type PageProps = { params: { id: string } };
