+++ src/app/blog/[post]/components/blog-post/blog-post.module.scss
