+++ src/app/blog/[post]/components/other-posts/other-post/other-post.tsx
