+++ src/app/category/components/category-hero/category-hero.module.scss
