+++ src/app/category/components/category-section/category-section.module.scss
