+++ src/app/category/page.tsx
@@ -1,5 +1,5 @@
-import { CategoryHero } from "@/components/category-hero/category-hero";
-import { CategorySection } from "@/components/category-section/category-section";
+import { CategoryHero } from "./components/category-hero/category-hero";
+import { CategorySection } from "./components/category-section/category-section";
 
 export default function Page() {
   return (
