+++ src/app/components/featured-posts/featured-post/featured-post.tsx
