+++ src/app/components/featured-posts/featured-posts.module.scss
