+++ src/app/components/logos-list/logos-list.module.scss
