+++ src/app/components/logos-list/logos-list.tsx
