+++ src/app/components/why-we-started/why-we-started.module.scss
