+++ src/app/layout.tsx
@@ -7,8 +7,8 @@ import { Header } from "@/components/header/header";
 import { inter, sen } from "@/styles/fonts";
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Client Blog",
+  description: "Modsen client blog Albedo-13",
 };
 
 export default function RootLayout({
